requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Collectfast==2.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-environ==0.9.0
django-extensions==3.2.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
django-storages==1.13.1
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
openapi-codec==1.3.2
phonenumbers==8.13.0
pytz==2022.6
requests==2.28.1
Shapely==1.8.5.post1
simplejson==3.18.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
